Custo e contato implementados

+++ src/main/java/com/br/pdvpostocombustivel/api/custo/dto/CustoRequest.java
@@ -2,16 +2,17 @@ package com.br.pdvpostocombustivel.api.custo.dto;
 
 import org.springframework.format.annotation.DateTimeFormat;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 /*
 Para Entrada
  */
 public record CustoRequest(
-        Double imposto,
-        Double custoVariavel,
-        Double custoFixo,
-        Double margemLucro,
+        BigDecimal imposto,
+        BigDecimal custoVariavel,
+        BigDecimal custoFixo,
+        BigDecimal margemLucro,
         @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate dataProcessamento
         )
 {}
